+++ src/strategy/EngineVolumeFirmYearSortStrategy.java
@@ -5,7 +5,7 @@ import models.Car;
 public class EngineVolumeFirmYearSortStrategy extends AbstractSortStrategy<Car> {
     @Override
     protected int compare(Car item1, Car item2) {
-        int result = Integer.compare(item1.getEngineVolume(), item2.getEngineVolume());
+        int result = Float.compare(item1.getEngineVolume(), item2.getEngineVolume());
         if (result == 0) {
             result = item1.getFirm().compareTo(item2.getFirm());
             if (result == 0) {
